Implemento correccion y agregado de cant me gusta, comentarios y si dio like en publicacion

+++ src/main/java/com/gymtrack/backend/dto/PublicacionDTO/PublicacionDTO.java
@@ -23,9 +23,9 @@ public class PublicacionDTO {
 
     private UsuarioResumenDTO usuario;
 
-    private Integer cantidadLikes;
+    private Long cantidadLikes;
 
-    private Integer cantidadComentarios;
+    private Long cantidadComentarios;
 
     private boolean dioLike; //si el usuario le dio like, el corazon cambiaria a rojo
